[!!!][TASK] Remove default parameters in step record
The default parameters in the step record are removed.
The parameters should be given explicitly in the form
finisher configuration for better transparency which
values are used. Also changing the parameters in the step
record can have unwanted side effects in a finisher if it
relies on the default parameters.